add rill yaml and restructure
