+++ web-local/src/routes/(files)/file/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+  import OnboardingPage from "../../(application)/+page.svelte";
+</script>
+
+<OnboardingPage />
