+++ web-local/src/routes/(files)/file/[...path]/[type=workspaces]/[name]/+page.svelte
