+++ web-local/src/routes/(files)/file/[...path]/dashboards/[name]/+page.svelte
