+++ web-local/src/routes/(files)/file/[...path]/rill.yaml/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+  import RillYamlPage from "../../../../(application)/rill.yaml/+page.svelte";
+</script>
+
+<RillYamlPage />
